refactor: extract MetricTile and TileDescriptor into their own file
